refactor: ColorPicker 컴포넌트 세분화 (#137)
- 기존 ColorPick에서 버튼과 파  팝업을 각 컴포넌트 별로 분리합니다.
